pyproject.toml
[tool.poetry]
name = "hrflow-connectors"
version = "4.23.0"
description = "hrflow-connectors is an open source project created by HrFlow.ai to allow developers to connect easily HR ecosystem component."
license = "Apache-2.0"
authors = ["HrFlow.ai <support+hrflow_connectors@hrflow.ai>"]
readme = "README.md"
homepage = "https://github.com/Riminder/hrflow-connectors"
repository = "https://github.com/Riminder/hrflow-connectors"
documentation = "https://github.com/Riminder/hrflow-connectors"
exclude = ["Makefile", ".pre-commit-config.yaml", "manifest.json", "tests", "poetry.lock"]

[tool.poetry.urls]
"HrFlow.ai Documentation" = "https://developers.hrflow.ai/"
"Changelog" = "https://github.com/Riminder/hrflow-connectors/releases/"


[tool.poetry.dependencies]
python = "^3.9"
pydantic = "1.10.8"
Jinja2 = "^3.0.3"
hrflow = "^4.2.0"
requests = "^2.27.1"
boto3 = {version = "^1.24.66", optional = true}
simple-salesforce = "^1.12.4"
typing-extensions = "^4.7.1"
beautifulsoup4 = "4.10.0"
puremagic = "^1.27"
msgspec = "^0.18.6"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
ipython = "^7.0"
flake8 = "^4.0.1"
flake8-isort = "^4.1.1"
flake8-black = "^0.3.6"
pre-commit = "^2.17.0"
pytest-cov = "^3.0.0"
ipdb = "^0.13.9"
nox = "^2022.11.21"
python-semantic-release = "^8.0.8"
pillow = "<10"
boto3-stubs = {extras = ["s3"], version = "^1.35.57"}

[tool.poetry.extras]
s3 = ["boto3"]

[tool.black]
line-length = 88
target-version = ["py39", "py310"]
preview = true
extend-exclude = ".*connector.pyi$"


[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.semantic_release]
commit_message = "{version}\n\nAutomatically generated by python-semantic-release"
commit_parser = "angular"
tag_format = "v{version}"
build_command = "poetry build"
version_toml = [
    "pyproject.toml:tool.poetry.version",
]

[tool.semantic_release.branches.main]
match = "(master)"
prerelease_token = "rc"
prerelease = false

[tool.semantic_release.commit_author]
env = "GIT_COMMIT_AUTHOR"
default = "hrflow-semantic-release <hrflow-semantic-release>"

[tool.semantic_release.commit_parser_options]
allowed_tags = ["build", "chore", "ci", "docs", "feat", "fix", "perf", "style", "refactor", "test"]
minor_tags = ["feat"]
patch_tags = ["fix", "perf"]

[tool.semantic_release.publish]
dist_glob_patterns = ["dist/*"]
upload_to_vcs_release = true
